requirements.txt
certifi
pika
python-dotenv
requests